feat(starship): character with λ, ɣ and ε
The character used in the prompt with lambda, seems very apt for an
anonymous function, which is each command I am sending. Then epsilon for
errors from the commands and gamma as the contrary to a lambda whenever
I am using vim-mode, which is very close to the vim symbol without being
the vim logo, which may be a good alternative as well.